requirements.txt
asn1crypto==0.24.0
atomicwrites==1.3.0
attrs==19.1.0
autobahn==19.3.3
Automat==0.7.0
certifi==2019.3.9
cffi==1.12.2
chardet==3.0.4
constantly==15.1.0
cryptography==2.6.1
cycler==0.10.0
dateparser==0.7.1
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
kiwisolver==1.0.1
matplotlib==3.0.3
more-itertools==7.0.0
numpy==1.16.2
pandas==0.24.2
pluggy==0.9.0
py==1.8.0
pyasn1==0.4.5
pyasn1-modules==0.2.4
pycparser==2.19
PyHamcrest==1.9.0
pyOpenSSL==19.0.0
pyparsing==2.4.0
pytest==4.4.0
python-binance==0.7.1
python-dateutil==2.8.0
pytz==2019.1
regex==2019.4.10
requests==2.21.0
service-identity==18.1.0
six==1.12.0
Twisted==19.2.0
txaio==18.8.1
tzlocal==1.5.1
urllib3==1.24.1
zope.interface==4.6.0
